all: move health and requestlog packages under server BREAKING_CHANGE_OK
