+++ gcp/gcpcloud/example_test.go
@@ -23,8 +23,8 @@ import (
 	"github.com/google/wire"
 	"go.opencensus.io/trace"
 	"gocloud.dev/gcp/gcpcloud"
-	"gocloud.dev/health"
 	"gocloud.dev/server"
+	"gocloud.dev/server/health"
 )
 
 // This is an example of how to bootstrap an HTTP server running on
