+++ internal/website/content/server/health/_index.md
@@ -1,0 +1,4 @@
+---
+title: gocloud.dev/server/health
+type: pkg
+---
