+++ internal/website/content/server/health/sqlhealth/_index.md
@@ -1,0 +1,4 @@
+---
+title: gocloud.dev/server/health/sqlhealth
+type: pkg
+---
