+++ internal/website/content/server/requestlog/_index.md
@@ -1,0 +1,4 @@
+---
+title: gocloud.dev/server/requestlog
+type: pkg
+---
