+++ samples/guestbook/wire_gen.go
@@ -26,13 +26,13 @@ import (
 	"gocloud.dev/gcp/cloudsql"
 	"gocloud.dev/mysql/awsmysql"
 	"gocloud.dev/mysql/gcpmysql"
-	"gocloud.dev/requestlog"
 	"gocloud.dev/runtimevar"
 	"gocloud.dev/runtimevar/awsparamstore"
 	"gocloud.dev/runtimevar/blobvar"
 	"gocloud.dev/runtimevar/filevar"
 	"gocloud.dev/runtimevar/gcpruntimeconfig"
 	"gocloud.dev/server"
+	"gocloud.dev/server/requestlog"
 	"gocloud.dev/server/sdserver"
 	"gocloud.dev/server/xrayserver"
 	"google.golang.org/genproto/googleapis/cloud/runtimeconfig/v1beta1"
