+++ server/health/sqlhealth/sqlhealth_test.go
@@ -22,7 +22,7 @@ import (
 	"sync"
 	"testing"
 
-	"gocloud.dev/health"
+	"gocloud.dev/server/health"
 )
 
 var _ = health.Checker((*Checker)(nil))
