+++ server/requestlog/requestlog.go
@@ -14,7 +14,7 @@
 
 // Package requestlog provides an http.Handler that logs information
 // about requests.
-package requestlog // import "gocloud.dev/requestlog"
+package requestlog // import "gocloud.dev/server/requestlog"
 
 import (
 	"errors"
